Fix MCP restructure test coverage
