+++ packages/shopstr-mcp/package.json
@@ -26,7 +26,7 @@
   "scripts": {
     "build": "tsc -p tsconfig.json",
     "start": "node dist/index.js",
-    "test": "npm run build && node --test test/*.node.mjs"
+    "test": "npm run build && node --test test/*.node.mjs test/tools/*.node.mjs"
   },
   "engines": {
     "node": ">=18"
